overloaded sql code
