+++ database/src/test/scala/za/co/absa/atum/database/runs/WriteCheckpointOverloadedIntegrationTests.scala
@@ -23,9 +23,9 @@ import za.co.absa.balta.classes.setter.CustomDBType
 import java.time.OffsetDateTime
 import java.util.UUID
 
-class WriteCheckpointV1IntegrationTests extends DBTestSuite {
+class WriteCheckpointOverloadedIntegrationTests extends DBTestSuite {
 
-  private val fnWriteCheckpointV1 = "runs.write_checkpoint_v1"
+  private val fnWriteCheckpoint = "runs.write_checkpoint"
 
   private val partitioning = JsonBString(
     """
@@ -41,6 +41,42 @@ class WriteCheckpointV1IntegrationTests extends DBTestSuite {
       |}
       |""".stripMargin
   )
+
+  private val measurements =
+    """
+      |{
+      |  "{
+      |    \"measure\": {
+      |      \"measureName\": \"count\",
+      |      \"measuredColumns\": []
+      |    },
+      |    \"result\":{
+      |      \"value\":\"3\",
+      |      \"type\":\"int\"
+      |    }
+      |  }",
+      |  "{
+      |    \"measure\": {
+      |      \"measureName\": \"avg\",
+      |      \"measuredColumns\": [\"col1\"]
+      |    },
+      |    \"result\":{
+      |      \"value\":\"3.14\",
+      |      \"type\":\"double\"
+      |    }
+      |  }",
+      |  "{
+      |    \"measure\": {
+      |      \"measureName\": \"avg\",
+      |      \"measuredColumns\": [\"a\",\"b\"]
+      |    },
+      |    \"result\":{
+      |      \"value\":\"2.71\",
+      |      \"type\":\"double\"
+      |    }
+      |  }"
+      |}
+      |""".stripMargin
 
   test("Write new checkpoint without data") {
     val uuid = UUID.randomUUID
@@ -64,8 +100,8 @@ class WriteCheckpointV1IntegrationTests extends DBTestSuite {
 
     assert(table("runs.checkpoints").count(add("fk_partitioning", fkPartitioning)) == 0)
 
-    function(fnWriteCheckpointV1)
-      .setParam("i_partitioning", partitioning)
+    function(fnWriteCheckpoint)
+      .setParam("i_partitioning_id", fkPartitioning)
       .setParam("i_id_checkpoint", uuid)
       .setParam("i_checkpoint_name", "Empty path")
       .setParam("i_process_start_time", startTime)
@@ -100,41 +136,7 @@ class WriteCheckpointV1IntegrationTests extends DBTestSuite {
     val user = "Franz Kafka"
     val startTime = OffsetDateTime.parse("1992-08-03T10:00:00Z")
     val endTime = OffsetDateTime.parse("2022-11-05T08:00:00Z")
-    val measurements =
-      """
-        |{
-        |  "{
-        |    \"measure\": {
-        |      \"measureName\": \"count\",
-        |      \"measuredColumns\": []
-        |    },
-        |    \"result\":{
-        |      \"value\":\"3\",
-        |      \"type\":\"int\"
-        |    }
-        |  }",
-        |  "{
-        |    \"measure\": {
-        |      \"measureName\": \"avg\",
-        |      \"measuredColumns\": [\"col1\"]
-        |    },
-        |    \"result\":{
-        |      \"value\":\"3.14\",
-        |      \"type\":\"double\"
-        |    }
-        |  }",
-        |  "{
-        |    \"measure\": {
-        |      \"measureName\": \"avg\",
-        |      \"measuredColumns\": [\"a\",\"b\"]
-        |    },
-        |    \"result\":{
-        |      \"value\":\"2.71\",
-        |      \"type\":\"double\"
-        |    }
-        |  }"
-        |}
-        |""".stripMargin
+
 
     table("runs.partitionings").insert(
       add("partitioning", partitioning)
@@ -152,8 +154,8 @@ class WriteCheckpointV1IntegrationTests extends DBTestSuite {
 
     assert(table("runs.checkpoints").count(add("fk_partitioning", fkPartitioning)) == 0)
 
-    function(fnWriteCheckpointV1)
-      .setParam("i_partitioning", partitioning)
+    function(fnWriteCheckpoint)
+      .setParam("i_partitioning_id", fkPartitioning)
       .setParam("i_id_checkpoint", uuid)
       .setParam("i_checkpoint_name", "Happy path")
       .setParam("i_process_start_time", startTime)
@@ -239,8 +241,8 @@ class WriteCheckpointV1IntegrationTests extends DBTestSuite {
         .add("created_by", origAuthor)
     )
 
-    function(fnWriteCheckpointV1)
-      .setParam("i_partitioning", partitioning)
+    function(fnWriteCheckpoint)
+      .setParam("i_partitioning_id", fkPartitioning)
       .setParam("i_id_checkpoint", uuid)
       .setParam("i_checkpoint_name", "Won't go in")
       .setParam("i_process_start_time", now())
@@ -266,8 +268,8 @@ class WriteCheckpointV1IntegrationTests extends DBTestSuite {
   test("Partitioning of the checkpoint does not exist") {
     val uuid = UUID.randomUUID
     val count = table("runs.checkpoints").count()
-    function(fnWriteCheckpointV1)
-      .setParam("i_partitioning", partitioning)
+    function(fnWriteCheckpoint)
+      .setParam("i_partitioning_id", 0L)
       .setParam("i_id_checkpoint", uuid)
       .setParam("i_checkpoint_name", "Won't go in")
       .setParam("i_process_start_time", now())
@@ -283,4 +285,5 @@ class WriteCheckpointV1IntegrationTests extends DBTestSuite {
       }
     assert(table("runs.checkpoints").count() == count)
   }
+
 }
