+++ server/src/main/scala/za/co/absa/atum/server/Main.scala
@@ -55,8 +55,8 @@ object Main extends ZIOAppDefault with Server {
           GetPartitioningAdditionalData.layer,
           CreateOrUpdateAdditionalData.layer,
           GetPartitioningCheckpoints.layer,
-          WriteCheckpointV1.layer,
           WriteCheckpoint.layer,
+          WriteCheckpointV2.layer,
           GetFlowCheckpoints.layer,
           PostgresDatabaseProvider.layer,
           TransactorProvider.layer,
