+++ server/src/main/scala/za/co/absa/atum/server/api/repository/CheckpointRepositoryImpl.scala
@@ -17,33 +17,33 @@
 package za.co.absa.atum.server.api.repository
 
 import za.co.absa.atum.model.dto.{CheckpointDTO, CheckpointV2DTO}
-import za.co.absa.atum.server.api.database.runs.functions.WriteCheckpoint.WriteCheckpointArgs
-import za.co.absa.atum.server.api.database.runs.functions.{WriteCheckpoint, WriteCheckpointV1}
+import za.co.absa.atum.server.api.database.runs.functions.WriteCheckpointV2.WriteCheckpointArgs
+import za.co.absa.atum.server.api.database.runs.functions.{WriteCheckpointV2, WriteCheckpoint}
 import za.co.absa.atum.server.api.exception.DatabaseError
 import zio._
 import zio.interop.catz.asyncInstance
 
-class CheckpointRepositoryImpl(writeCheckpointV1Fn: WriteCheckpointV1, writeCheckpointFn: WriteCheckpoint)
+class CheckpointRepositoryImpl(writeCheckpointFn: WriteCheckpoint, writeCheckpointV2Fn: WriteCheckpointV2)
     extends CheckpointRepository
     with BaseRepository {
 
   override def writeCheckpoint(checkpointDTO: CheckpointDTO): IO[DatabaseError, Unit] = {
-    dbSingleResultCallWithStatus(writeCheckpointV1Fn(checkpointDTO), "writeCheckpoint")
+    dbSingleResultCallWithStatus(writeCheckpointFn(checkpointDTO), "writeCheckpoint")
   }
 
   override def writeCheckpointV2(partitioningId: Long, checkpointV2DTO: CheckpointV2DTO): IO[DatabaseError, Unit] = {
     dbSingleResultCallWithStatus(
-      writeCheckpointFn(WriteCheckpointArgs(partitioningId, checkpointV2DTO)),
+      writeCheckpointV2Fn(WriteCheckpointArgs(partitioningId, checkpointV2DTO)),
       "writeCheckpoint"
     )
   }
 }
 
 object CheckpointRepositoryImpl {
-  val layer: URLayer[WriteCheckpointV1 with WriteCheckpoint, CheckpointRepository] = ZLayer {
+  val layer: URLayer[WriteCheckpoint with WriteCheckpointV2, CheckpointRepository] = ZLayer {
     for {
-      writeCheckpointV1 <- ZIO.service[WriteCheckpointV1]
       writeCheckpoint <- ZIO.service[WriteCheckpoint]
-    } yield new CheckpointRepositoryImpl(writeCheckpointV1, writeCheckpoint)
+      writeCheckpointV2 <- ZIO.service[WriteCheckpointV2]
+    } yield new CheckpointRepositoryImpl(writeCheckpoint, writeCheckpointV2)
   }
 }
