+++ server/src/test/scala/za/co/absa/atum/server/api/database/runs/functions/WriteCheckpointIntegrationTests.scala
@@ -17,14 +17,13 @@
 package za.co.absa.atum.server.api.database.runs.functions
 
 import za.co.absa.atum.model.ResultValueType
-import za.co.absa.atum.model.dto._
+import za.co.absa.atum.model.dto.{CheckpointDTO, MeasureDTO, MeasureResultDTO, MeasurementDTO, PartitionDTO}
 import za.co.absa.atum.model.dto.MeasureResultDTO.TypedValue
 import za.co.absa.atum.server.ConfigProviderTest
 import za.co.absa.atum.server.api.TestTransactorProvider
 import za.co.absa.atum.server.api.database.PostgresDatabaseProvider
-import za.co.absa.atum.server.api.database.runs.functions.WriteCheckpoint.WriteCheckpointArgs
-import za.co.absa.db.fadb.exceptions.DataConflictException
-import za.co.absa.db.fadb.status.{FunctionStatus, Row}
+import za.co.absa.db.fadb.exceptions.{DataConflictException, DataNotFoundException}
+import za.co.absa.db.fadb.status.FunctionStatus
 import zio._
 import zio.interop.catz.asyncInstance
 import zio.test._
@@ -36,12 +35,13 @@ object WriteCheckpointIntegrationTests extends ConfigProviderTest {
 
   override def spec: Spec[TestEnvironment with Scope, Any] = {
 
-    suite("WriteCheckpointV2Suite")(
+    suite("WriteCheckpointSuite")(
       test("Returns expected Left with DataNotFoundException as related partitioning is not in the database") {
-        val checkpointV2DTO = CheckpointV2DTO(
+        val checkpointDTO = CheckpointDTO(
           id = UUID.randomUUID(),
           name = "name",
           author = "author",
+          partitioning = Seq(PartitionDTO("key4", "value4")),
           processStartTime = ZonedDateTime.now(),
           processEndTime = Option(ZonedDateTime.now()),
           measurements = Set(
@@ -49,8 +49,8 @@ object WriteCheckpointIntegrationTests extends ConfigProviderTest {
           )
         )
         for {
-          writeCheckpointV2 <- ZIO.service[WriteCheckpoint]
-          result <- writeCheckpointV2(WriteCheckpointArgs(1L, checkpointV2DTO))
+          writeCheckpoint <- ZIO.service[WriteCheckpoint]
+          result <- writeCheckpoint(checkpointDTO)
         } yield assertTrue(result == Left(DataConflictException(FunctionStatus(32, "Partitioning not found"))))
       }
     ).provide(
