+++ server/src/test/scala/za/co/absa/atum/server/api/database/runs/functions/WriteCheckpointV2IntegrationTests.scala
@@ -17,13 +17,14 @@
 package za.co.absa.atum.server.api.database.runs.functions
 
 import za.co.absa.atum.model.ResultValueType
-import za.co.absa.atum.model.dto.{CheckpointDTO, MeasureDTO, MeasureResultDTO, MeasurementDTO, PartitionDTO}
+import za.co.absa.atum.model.dto._
 import za.co.absa.atum.model.dto.MeasureResultDTO.TypedValue
 import za.co.absa.atum.server.ConfigProviderTest
 import za.co.absa.atum.server.api.TestTransactorProvider
 import za.co.absa.atum.server.api.database.PostgresDatabaseProvider
-import za.co.absa.db.fadb.exceptions.{DataConflictException, DataNotFoundException}
-import za.co.absa.db.fadb.status.FunctionStatus
+import za.co.absa.atum.server.api.database.runs.functions.WriteCheckpointV2.WriteCheckpointArgs
+import za.co.absa.db.fadb.exceptions.DataConflictException
+import za.co.absa.db.fadb.status.{FunctionStatus, Row}
 import zio._
 import zio.interop.catz.asyncInstance
 import zio.test._
@@ -31,17 +32,16 @@ import zio.test._
 import java.time.ZonedDateTime
 import java.util.UUID
 
-object WriteCheckpointV1IntegrationTests extends ConfigProviderTest {
+object WriteCheckpointV2IntegrationTests extends ConfigProviderTest {
 
   override def spec: Spec[TestEnvironment with Scope, Any] = {
 
-    suite("WriteCheckpointSuite")(
+    suite("WriteCheckpointV2Suite")(
       test("Returns expected Left with DataNotFoundException as related partitioning is not in the database") {
-        val checkpointDTO = CheckpointDTO(
+        val checkpointV2DTO = CheckpointV2DTO(
           id = UUID.randomUUID(),
           name = "name",
           author = "author",
-          partitioning = Seq(PartitionDTO("key4", "value4")),
           processStartTime = ZonedDateTime.now(),
           processEndTime = Option(ZonedDateTime.now()),
           measurements = Set(
@@ -49,12 +49,12 @@ object WriteCheckpointV1IntegrationTests extends ConfigProviderTest {
           )
         )
         for {
-          writeCheckpoint <- ZIO.service[WriteCheckpointV1]
-          result <- writeCheckpoint(checkpointDTO)
+          writeCheckpointV2 <- ZIO.service[WriteCheckpointV2]
+          result <- writeCheckpointV2(WriteCheckpointArgs(1L, checkpointV2DTO))
         } yield assertTrue(result == Left(DataConflictException(FunctionStatus(32, "Partitioning not found"))))
       }
     ).provide(
-      WriteCheckpointV1.layer,
+      WriteCheckpointV2.layer,
       PostgresDatabaseProvider.layer,
       TestTransactorProvider.layerWithRollback
     )
