+++ server/src/test/scala/za/co/absa/atum/server/api/repository/CheckpointRepositoryUnitTests.scala
@@ -17,10 +17,10 @@
 package za.co.absa.atum.server.api.repository
 
 import org.mockito.Mockito.{mock, when}
-import za.co.absa.atum.server.api.database.runs.functions.{WriteCheckpoint, WriteCheckpointV1}
+import za.co.absa.atum.server.api.database.runs.functions.{WriteCheckpointV2, WriteCheckpoint}
 import za.co.absa.atum.server.api.exception.DatabaseError
 import za.co.absa.atum.server.api.TestData
-import za.co.absa.atum.server.api.database.runs.functions.WriteCheckpoint.WriteCheckpointArgs
+import za.co.absa.atum.server.api.database.runs.functions.WriteCheckpointV2.WriteCheckpointArgs
 import za.co.absa.atum.server.api.exception.DatabaseError._
 import za.co.absa.db.fadb.exceptions.DataConflictException
 import za.co.absa.db.fadb.status.FunctionStatus
@@ -32,25 +32,25 @@ import za.co.absa.db.fadb.status.Row
 
 object CheckpointRepositoryUnitTests extends ZIOSpecDefault with TestData {
 
-  private val writeCheckpointV1Mock: WriteCheckpointV1 = mock(classOf[WriteCheckpointV1])
   private val writeCheckpointMock: WriteCheckpoint = mock(classOf[WriteCheckpoint])
+  private val writeCheckpointV2Mock: WriteCheckpointV2 = mock(classOf[WriteCheckpointV2])
 
-  when(writeCheckpointV1Mock.apply(checkpointDTO1)).thenReturn(ZIO.right(Row(FunctionStatus(0, "success"), ())))
-  when(writeCheckpointV1Mock.apply(checkpointDTO2))
+  when(writeCheckpointMock.apply(checkpointDTO1)).thenReturn(ZIO.right(Row(FunctionStatus(0, "success"), ())))
+  when(writeCheckpointMock.apply(checkpointDTO2))
     .thenReturn(ZIO.fail(DataConflictException(FunctionStatus(31, "conflict"))))
-  when(writeCheckpointV1Mock.apply(checkpointDTO3)).thenReturn(ZIO.fail(new Exception("boom!")))
+  when(writeCheckpointMock.apply(checkpointDTO3)).thenReturn(ZIO.fail(new Exception("boom!")))
 
   private val partitioningId = 1L
 
-  when(writeCheckpointMock.apply(WriteCheckpointArgs(partitioningId, checkpointV2DTO1)))
+  when(writeCheckpointV2Mock.apply(WriteCheckpointArgs(partitioningId, checkpointV2DTO1)))
     .thenReturn(ZIO.right(Row(FunctionStatus(0, "success"), ())))
-  when(writeCheckpointMock.apply(WriteCheckpointArgs(partitioningId, checkpointV2DTO2)))
+  when(writeCheckpointV2Mock.apply(WriteCheckpointArgs(partitioningId, checkpointV2DTO2)))
     .thenReturn(ZIO.left(DataConflictException(FunctionStatus(32, "Partitioning not found"))))
-  when(writeCheckpointMock.apply(WriteCheckpointArgs(partitioningId, checkpointV2DTO3)))
+  when(writeCheckpointV2Mock.apply(WriteCheckpointArgs(partitioningId, checkpointV2DTO3)))
     .thenReturn(ZIO.fail(new Exception("boom!")))
 
-  private val writeCheckpointV1MockLayer = ZLayer.succeed(writeCheckpointV1Mock)
   private val writeCheckpointMockLayer = ZLayer.succeed(writeCheckpointMock)
+  private val writeCheckpointV2MockLayer = ZLayer.succeed(writeCheckpointV2Mock)
 
   override def spec: Spec[TestEnvironment with Scope, Any] = {
 
@@ -87,7 +87,7 @@ object CheckpointRepositoryUnitTests extends ZIOSpecDefault with TestData {
           )
         }
       )
-    ).provide(CheckpointRepositoryImpl.layer, writeCheckpointV1MockLayer, writeCheckpointMockLayer)
+    ).provide(CheckpointRepositoryImpl.layer, writeCheckpointMockLayer, writeCheckpointV2MockLayer)
 
   }
